Setup player animations
Added animations for all kinds of movement in the game so far

All of them are controlled via the Animator window.

Parameters are setup in the animator window and those are used by transitions to determine when to play an animation.
The parameters get their values from the PlayerMovement script.