Add gun to player
Make a pistol object and attach it to the player as a child object.

The flip function of the character has also been modified so that the whole gameobject (including children) filps instead of just the sprite.

Eventually the weapon child object should just be some generic object that we can assign guns to so new weapons can be picked up eventually.